Work around AngularJS stupidity
Change all pseudo-self-closing-tags to explicitly closed tags

AngularJS is a really *clever* framework. But good code should *not* be clever,
just be easily understandable and easy to learn.

AngularJS tries hard to make you think what you are writing is HTML. It is
*not*. You will believe it really long, and one day it will bite you.

AngularJS creates custom HTML tags. How could anything ever possibly go wrong?
The specification doesn't say much about unknown elements, and voila, you're
hurt.

Here, the issue is that HTML says elements cannot be self-closing, except for a
tiny number of enumerated tags. And AngularJS' tags of course are not in the set
of what can be self-closed. So, the natural effect is that they are not closed,
and immediately englobe all the elements until their parent is closed.

Does that sound normal? AngularJS devs say this is not an AngularJS bug. They
are right, *this* is not an AngularJS bug. The issue is that AngularJS even
tries to make you believe what you are writing is normal html. Let me hammer it
down once more, *it* *is* *not*.

AngularJS is fundamentally broken.

It hacks on a specification that is not designed to handle it. Undoubtedly, it
is a really amazing lot of clever code, but it is not usable. As soon as you try
thinking as they want you to think, that is as though you are writnig "normal"
HTML with special tags, the game is over and you discover it is not.

What you are writing is not HTML, it is AngularJS' pseudolanguage. And it looks
like it is designed to break on every (x)html version change: it draws so deeply
on the internals of the language that any non-conformance will simply make it
not work.

If they wanted to make their own pseudolanguage, they really should have
actually used one, and even thrown xml out of the window. Anyway, xml is a
failure too in modern networks, with its redundant verbose tags. They could just
have written a parser in javascript and used it -- anyway, when you use such a
huge framework you are *expecting* the webpage to be extremely slow to load.

So AngularJS developers chose the worst of the options: they relied on HTML
parsing by the browser, but at the same time decided not to use standard
elements, thus creating a domain-specific language. THEY CREATED A DSL WITH A
PARSER THEY DID NOT WRITE. Yes, you read it right, it's completely stupid: any
bug in any browser's parser will just make the website unreadable. And it's not
like it's even often used code paths. No, they chose to use the least normal
code paths, by deciding to use elements that actually do not exist and therefore
do not have the same nth-class-citizenship as the others.

So, yes, it's clever, but actually such clever hacks should only be used during
the IOCCC, not when writing production code.

I think I'm done for now, but to recap:

AngularJS' fundamental design is broken due to cleverness